requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
firebase-admin==6.6.0
fuzzywuzzy==0.18.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
Levenshtein==0.26.1
libmagic==1.0
motor==3.7.0
msgpack==1.1.0
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.9
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
PyMySQL==1.1.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-Levenshtein==0.26.1
python-magic==0.4.27
python-multipart==0.0.6
qrcode==8.0
RapidFuzz==3.12.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==10.4
pydantic_settings
